Document environment variable for OpenAI
