+++ README.md
@@ -11,3 +11,11 @@ npm run dev
 ```
 
 Open `http://localhost:5173` in your browser to view the application.
+
+## Environment Variables
+
+Create a `.env` file inside `my-app` with your OpenAI key:
+
+```
+VITE_OPENAI_API_KEY=your-api-key
+```
